Refactor(renderer): Move text-expression binding into blocks/expression.js
