+++ packages/renderer/src/engines/native/reactive-data.js
@@ -1,5 +1,4 @@
 import { isArray, isFunction, isPlainObject } from '@semantic-ui/utils';
-import { isBlockClose, isBlockOpen, isExpressionMarker } from '../../build-html-string.js';
 import { renderASTToString } from './commit-hooks.js';
 
 // Evaluate a marker entry to its attribute-value string. Expression entries
@@ -142,113 +141,6 @@ export function bindAttribute({
   }
 }
 
-// Fresh text-expression binding. Three shapes:
-// • unsafeHTML  — parse value as HTML, replace owned nodes on each run
-// • literalValue — static text node, no reaction
-// • default      — reactive text node, update .data on each run
-export function bindTextExpression({ comment, entry, data, scope, renderer }) {
-  const exprNode = entry.node;
-  const parent = comment.parentNode;
-
-  if (exprNode.unsafeHTML) {
-    const anchor = document.createTextNode('');
-    comment.replaceWith(anchor);
-    const ownedNodes = [];
-    scope.reaction(anchor, () => {
-      for (const n of ownedNodes) { n.remove(); }
-      ownedNodes.length = 0;
-      const value = renderer.lookupExpression(exprNode.value, data);
-      if (value != null && value !== '') {
-        const parsed = renderer.parseHTML(String(value));
-        const nodes = [...parsed.childNodes];
-        anchor.after(parsed);
-        ownedNodes.push(...nodes);
-      }
-    });
-  }
-  else if (exprNode.literalValue) {
-    const value = renderer.evaluator.lookupTokenValue(exprNode.value, data);
-    const textNode = document.createTextNode(value ?? '');
-    parent.replaceChild(textNode, comment);
-  }
-  else {
-    const textNode = document.createTextNode('');
-    parent.replaceChild(textNode, comment);
-    scope.reaction(textNode, () => {
-      const value = renderer.lookupExpression(exprNode.value, data);
-      textNode.data = value ?? '';
-    });
-  }
-}
-
-// Hydrating text-expression binding — adopts server-rendered DOM instead
-// of creating fresh nodes. The server output merges the value with any
-// following static text into a single text node; we split it at the
-// server-value boundary and adopt the first part as the reactive node.
-export function hydrateTextExpression({ comment, entry, data, scope, renderer }) {
-  const exprNode = entry.node;
-
-  if (exprNode.unsafeHTML) {
-    // Collect server-rendered nodes after the comment until next marker
-    const ownedNodes = [];
-    let next = comment.nextSibling;
-    while (
-      next && !(next.nodeType === Node.COMMENT_NODE
-        && (isExpressionMarker(next.data) || isBlockOpen(next.data) || isBlockClose(next.data)))
-    ) {
-      ownedNodes.push(next);
-      next = next.nextSibling;
-    }
-
-    const anchor = document.createTextNode('');
-    comment.replaceWith(anchor);
-
-    scope.reaction(anchor, (comp) => {
-      const value = renderer.lookupExpression(exprNode.value, data);
-      if (comp.firstRun) { return; } // skip expensive reparse — server DOM is trusted
-      for (const n of ownedNodes) { n.remove(); }
-      ownedNodes.length = 0;
-      if (value != null && value !== '') {
-        const parsed = renderer.parseHTML(String(value));
-        const nodes = [...parsed.childNodes];
-        anchor.after(parsed);
-        ownedNodes.push(...nodes);
-      }
-    });
-    return;
-  }
-
-  // Safe text expression — the browser merges VALUE with any following
-  // static text into one text node. Split at the boundary so the reactive
-  // text node only covers the value portion.
-  const nextNode = comment.nextSibling;
-  let textNode;
-  if (nextNode && nextNode.nodeType === Node.TEXT_NODE) {
-    const serverValue = String(renderer.lookupExpression(exprNode.value, data) ?? '');
-    const fullText = nextNode.data;
-    if (fullText.length > serverValue.length && fullText.startsWith(serverValue)) {
-      nextNode.splitText(serverValue.length);
-      textNode = nextNode;
-    }
-    else {
-      textNode = nextNode;
-    }
-    comment.remove();
-  }
-  else {
-    textNode = document.createTextNode('');
-    comment.replaceWith(textNode);
-  }
-
-  scope.reaction(textNode, (comp) => {
-    if (comp.firstRun) {
-      renderer.lookupExpression(exprNode.value, data);
-      return;
-    }
-    const value = renderer.lookupExpression(exprNode.value, data);
-    textNode.data = value ?? '';
-  });
-}
-
-// Raw-text element binding (script, style, textarea, title) moved to
-// blocks/raw-text.js. Dispatched via the block registry as `'rawText'`.
+// Text-position expression bindings (fresh + hydrating) moved to
+// blocks/expression.js. Raw-text bindings moved to blocks/raw-text.js.
+// Both dispatched via the block registry as `'expression'` / `'rawText'`.
